requirements.txt
flask
scikit-learn
joblib
pillow
numpy
scipy
